Move adapters into models
