+++ src/core/models/filter.js
@@ -39,6 +39,20 @@ export default class Filter {
   }
 
   /**
+   * Constructs an SDK Filter model from an answers-core SimpleFilter model
+   *
+   * @param {SimpleFilter} filter from answers-core
+   * @returns {@link Filter}
+   */
+  static fromCoreSimpleFilter (filter) {
+    return new Filter({
+      [filter.fieldId]: {
+        [filter.comparator]: filter.comparedValue
+      }
+    });
+  }
+
+  /**
    * Parse a JSON format filter returned from the server into a Filter
    * @param {*} responseFilter A filter in JSON format returned from the backend
    * @returns {Filter}
