+++ src/core/models/result.js
@@ -1,5 +1,8 @@
 /** @module Result */
 
+import HighlightedFieldMap from './highlightedfieldmap';
+import { truncate } from '../utils/strings';
+
 export default class Result {
   constructor (data = {}) {
     /**
@@ -100,5 +103,29 @@ export default class Result {
      * @type {number}
      */
     this.distanceFromFilter = data.distanceFromFilter || null;
+  }
+
+  /**
+   * Constructs an SDK Result from an answers-core Result
+   *
+   * @param {Result} result from answers-core
+   * @returns {@link Result}
+   */
+  static fromCore (result) {
+    const highlightedData = HighlightedFieldMap.fromCore(result.highlightedValues);
+    const details = highlightedData.description || result.description;
+    const truncatedDetails = truncate(details);
+
+    return new Result({
+      raw: result.rawData,
+      ordinal: result.index,
+      title: result.name,
+      details: truncatedDetails,
+      link: result.link,
+      id: result.id,
+      distance: result.distance,
+      distanceFromFilter: result.distanceFromFilter,
+      highlighted: highlightedData
+    });
   }
 }
